Temporary fix to movie search
