+++ Selenium Lab/website/__init__.py
@@ -6,8 +6,10 @@ def create_app():
 
     from .views import views
     from .auth import auth
+    from .movies import movie
 
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(auth, url_prefix='/')
+    app.register_blueprint(movie, url_prefix="/")
 
     return app
